fix do not delete oldprogram and do not import if already imported

+++ app/Test/data/mongoDB/dropAll.js
@@ -2,7 +2,9 @@ var myDatabases = db.adminCommand('listDatabases')
 
 for (var i=0; i < myDatabases['databases'].length; i++)
 {
-	db = db.getMongo().getDB(myDatabases['databases'][i]['name'])
-	print('Dropping: '+db.getName());
-	db.dropDatabase();
+    if (myDatabases['databases'][i]['name'] != 'oldprogram') {
+        db = db.getMongo().getDB(myDatabases['databases'][i]['name'])
+        print('Dropping: '+db.getName());
+        db.dropDatabase();
+    }
 }
